image-crawler/requirements.txt
requests>=2.32.3
pillow>=10.4.0
opencv-python>=4.10.0.84
